+++ Proyecto SCADA/scada-front/scada-front/src/app/pages/sinopticos/sinoptico-editor/store/editor-store.ts
@@ -125,6 +125,7 @@ export class EditorStore {
       link: { width: 160, height: 50 },
       clock: { width: 200, height: 100 },
       variable: { width: 240, height: 120 },
+      alarm: { width: 350, height: 250 },
     };
     const size = defaults[type];
     const widget: CanvasWidget = {